+++ src/kernel/mem.h
@@ -43,7 +43,7 @@ int memcmp(const void* lhs, const void *rhs, size_t size);
 
 int strcmp(const char* s1, const char* s2);
 int strcpy(char *dst, const char *src);
-int strlen(const char *str);
+size_t strlen(const char *str);
 char* strtok(char *s, const char *delim);
 
 // Page X table entry
